fix warnings with MSVC

+++ WDL/IPlug/IPlugVST.cpp
@@ -201,7 +201,7 @@ EHost IPlugVST::GetHost()
 
     if (CSTR_NOT_EMPTY(productStr))
     {
-      int decVer = mHostCallback(&mAEffect, audioMasterGetVendorVersion, 0, 0, 0, 0.0f);
+      int decVer = (int)mHostCallback(&mAEffect, audioMasterGetVendorVersion, 0, 0, 0, 0.0f);
       int ver = decVer / 10000;
       int rmaj = (decVer - 10000 * ver) / 100;
       int rmin = (decVer - 10000 * ver - 100 * rmaj);
@@ -418,7 +418,7 @@ VstIntPtr VSTCALLBACK IPlugVST::VSTDispatcher(AEffect *pEffect, VstInt32 opCode,
     }
     case effSetBlockSize:
     {
-      _this->SetBlockSize(value);
+      _this->SetBlockSize((int)value);
       _this->Reset();
       return 0;
     }
@@ -518,7 +518,7 @@ VstIntPtr VSTCALLBACK IPlugVST::VSTDispatcher(AEffect *pEffect, VstInt32 opCode,
       {
         bool isBank = (!idx);
         ByteChunk* pChunk = (isBank ? &(_this->mBankState) : &(_this->mState));
-        pChunk->Resize(value);
+        pChunk->Resize((int)value);
         memcpy(pChunk->GetBytes(), ptr, value);
         int pos = 0;
         int iplugVer = _this->GetIPlugVerFromChunk(pChunk, &pos);
@@ -733,7 +733,7 @@ VstIntPtr VSTCALLBACK IPlugVST::VSTDispatcher(AEffect *pEffect, VstInt32 opCode,
       {
         if (value >= 0 && value < _this->NParams())
         {
-          _this->GetParam(value)->GetDisplayForHost((double) opt, true, (char*) ptr);
+          _this->GetParam((int)value)->GetDisplayForHost((double) opt, true, (char*) ptr);
         }
         return 0xbeef;
       }
@@ -742,7 +742,7 @@ VstIntPtr VSTCALLBACK IPlugVST::VSTDispatcher(AEffect *pEffect, VstInt32 opCode,
       {
         if (value >= 0 && value < _this->NParams())
         {
-          if (_this->GetParam(value)->Type() != IParam::kTypeDouble)
+          if (_this->GetParam((int)value)->Type() != IParam::kTypeDouble)
           {
             return 0xbeef;
           }
